+++ source/ChanSort.Loader.Hisense2017/HisDbSerializer.cs
@@ -357,10 +357,7 @@ left outer join Lcn l on l.ServiceId=fi.ServiceId and l.FavoriteId=fi.FavoriteId
           
           int favListIdx = favListIdToFavIndex.TryGet(favListId, -1);
           if (favListIdx >= 0)
-          {
-            // NOTE: we need to set the NEW fav index here because AddChannel will use the new value to initialize the old value
-            ci.FavIndex[favListIdx] = r.GetInt32(2);
-          }
+            ci.OldFavIndex[favListIdx] = r.GetInt32(2);
 
           ci.SetOldPosition(favListIdx + 1, r.GetInt32(2)); // 0=main nr, 1-4=fav 1-4
           if (favListIdx < 0)